requirements.txt
discord-py==2.6.4
pydantic==2.12.3
pydantic-settings==2.11.0
pynacl==1.6.0
